user dashboard / persistent events
